+++ src/main/kotlin/nl/devhaan/kotlinpoetdsl/codeblock/LazyComponentAcceptor.kt
@@ -1,11 +1,14 @@
 package nl.devhaan.kotlinpoetdsl.codeblock
 
+import com.squareup.kotlinpoet.CodeBlock
 import nl.devhaan.kotlinpoetdsl.CodeBlockLevel
 import nl.devhaan.kotlinpoetdsl.helpers.BlockWrapper
+import nl.devhaan.kotlinpoetdsl.helpers.wrapper
 
 @CodeBlockLevel
 interface LazyComponentAcceptor
 
-class LazyComponent(val wrapper : (BlockWrapper<*, *, *>) -> Unit)
+class LazyComponent(val wrapper: (BlockWrapper<*, *, *>) -> Unit, val singleStatement: Boolean)
 
-fun createLazyComponent(builder : (BlockWrapper<*, *, *>) ->Unit) = LazyComponent(builder)
+fun LazyComponent.toCodeBlock() = CodeBlock.builder().wrapper().also(wrapper).build()
+fun createLazyComponent(singleStatement: Boolean, builder: (BlockWrapper<*, *, *>) -> Unit) = LazyComponent(builder, singleStatement)